+++ packages/core/src/internal/process/transform/convertDeploymentVertexToExecutionVertex.ts
@@ -6,7 +6,6 @@ import {
   EventParamFuture,
 } from "../../../types/future";
 import { Artifact } from "../../../types/hardhat";
-import { Services } from "../../../types/services";
 import { IgnitionError } from "../../../utils/errors";
 import { isBytesArg, isFuture } from "../../../utils/guards";
 import {
@@ -31,6 +30,7 @@ import {
   LibraryDeploy,
   SentETH,
 } from "../../types/executionGraph";
+import { Services } from "../../types/services";
 
 interface TransformContext {
   services: Services;
