+++ packages/core/src/internal/services/createServices.ts
@@ -1,8 +1,8 @@
 import { ethers } from "ethers";
 
-import { Providers } from "../types/providers";
+import { Providers } from "../../types/providers";
+import { TxSender } from "../../utils/tx-sender";
 import { Services } from "../types/services";
-import { TxSender } from "../utils/tx-sender";
 
 import { AccountsService } from "./AccountsService";
 import { ArtifactsService } from "./ArtifactsService";
