+++ packages/core/test/execution/dispatch.ts
@@ -9,8 +9,11 @@ import { execute } from "../../src/internal/execution/execute";
 import { ExecutionVertex } from "../../src/internal/types/executionGraph";
 import { VertexResultEnum } from "../../src/internal/types/graph";
 import { ICommandJournal } from "../../src/internal/types/journal";
+import {
+  Services,
+  TransactionOptions,
+} from "../../src/internal/types/services";
 import { Artifact } from "../../src/types/hardhat";
-import { Services, TransactionOptions } from "../../src/types/services";
 import { buildAdjacencyListFrom } from "../graph/helpers";
 import { getMockServices } from "../helpers";
